C1/requirements.txt
requests==2.20.0
requests_futures==0.9.8
urllib3==1.23
python-dateutil==1.4
